moved all of lucy's code to one file
