+++ app/build.gradle
@@ -27,4 +27,6 @@ dependencies {
     compile 'io.indico:android:1.0.2'
     compile 'com.mcxiaoke.volley:library:1.0.19'
     compile 'com.jakewharton:butterknife:7.0.1'
+    compile 'org.apache.commons:commons-lang3:3.0'
+
 }
