+++ app/src/main/java/com/mobileproto/hireddit/hireddit/reddit/CommentCallback.java
@@ -6,5 +6,5 @@ import java.util.ArrayList;
  * Created by lwilcox on 11/5/2015.
  */
 public interface CommentCallback {
-    void callback(ArrayList<String> commentList);
+    void commentCallback(String comment);
 }
