+++ app/src/main/java/com/mobileproto/hireddit/hireddit/reddit/RedditSearchCallback.java
@@ -1,0 +1,11 @@
+package com.mobileproto.hireddit.hireddit.reddit;
+
+import java.util.ArrayList;
+
+/**
+ * Created by david on 11/17/15.
+ */
+public interface RedditSearchCallback
+{
+    void callback(ArrayList<String> comments);
+}
