+++ app/src/main/java/com/mobileproto/hireddit/hireddit/visuals/SpeechFragment.java
@@ -16,7 +16,8 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.mobileproto.hireddit.hireddit.R;
-import com.mobileproto.hireddit.hireddit.reddit.GetWordsAsync;
+import com.mobileproto.hireddit.hireddit.reddit.CommentCallback;
+import com.mobileproto.hireddit.hireddit.reddit.RedditSearcher;
 import com.mobileproto.hireddit.hireddit.speech.SpeechCallback;
 import com.mobileproto.hireddit.hireddit.speech.SpeechListener;
 
@@ -33,7 +34,8 @@ import butterknife.ButterKnife;
  * Use the {@link SpeechFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class SpeechFragment extends Fragment implements SpeechCallback {
+public class SpeechFragment extends Fragment implements SpeechCallback, CommentCallback
+{
     private OnFragmentInteractionListener mListener;
     private static final String DEBUG_TAG = "SpeechFragment Debug";
     private boolean isListening;
@@ -144,7 +146,7 @@ public class SpeechFragment extends Fragment implements SpeechCallback {
         voiceInput = voiceResult;
         String firstResult = voiceInput.get(0).toString();
         speechTextDisplay.setText(firstResult);
-        new GetWordsAsync(firstResult, getActivity().getApplicationContext(), commentText).execute();
+        new RedditSearcher(firstResult, getActivity().getApplicationContext(), commentText).execute();
         isListening = false;
         Log.d(DEBUG_TAG, "Got result, stopped listening.");
     }
@@ -172,6 +174,13 @@ public class SpeechFragment extends Fragment implements SpeechCallback {
             }
         }
     }
+
+    @Override
+    public void commentCallback(String comment)
+    {
+        commentText.setText(comment);
+    }
+
 
     /**
      * This interface must be implemented by activities that contain this
